requirements.txt
flask
gunicorn
psycopg2-binary  # Für PostgreSQL
python-dotenv     # Für Umgebungsvariablen (optional)